user login

+++ database/migrations/2026_04_29_155221_create_tarif_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tarif', function (Blueprint $table) {
             $table->id();
             $table->integer('daya');
-            $table->integer('tarifperkwh');
+            $table->decimal('tarifperkwh', 15, 2);
             $table->timestamps();
         });
     }
